fix(map): unable to setup test infrastructure
The setup of certificates to remove during the test need to use
`map.jinja` but it fails to load the YAML files since the `tplroot` is
not the `cert` formula directory in this case.

We need to enable the import of `mapdata` from outside the
`cert-formula` since `tplroot` can't be set correctly when the import:

- is done from another top directory than the `cert`
- the import is done `with context`

In this case, the `tpldir` is set to the directory of the importer
`.sls` file instead of the `.jinja` imported one.

We force the `without context` which permits to directly use `tpldir`
as the `tplroot` which is the directory of the imported file.

BREAKING CHANGE: `map.jinja` import must use `without context`

BREAKING CHANGE: `libmapstack.jinja` import must use `without context`

BREAKING CHANGE: `libmatchers.jinja` import must use `without context`